Refactor library hooks: separate useAuthors and useBooks, add tests for new hooks
