+++ src/hooks/useBooks/utils.test.ts
@@ -1,21 +1,13 @@
 import { cleanup } from "@testing-library/react";
 import { describe, afterEach, expect } from "vitest";
-import { add_author_by_name, add_book_by_title_and_author_id } from "./utils";
 import { library_test } from "../../testing_utils";
+import { add_book_by_title_and_author_id } from "./utils";
 
-describe("use_library", () => {
+describe("useBooks utils", () => {
   afterEach(() => {
     cleanup();
   });
-  library_test("add_author_by_name util", ({ authors }) => {
-    const mock_author_name = "JRR Mockien";
-    const new_authors = add_author_by_name(authors, mock_author_name);
-
-    expect(new_authors.length).toBe(authors.length + 1);
-    expect(new_authors[2].name).toBe(mock_author_name);
-  });
-
-  library_test("add_book_by_title_and_author_id util", ({ books }) => {
+  library_test("add_book_by_title_and_author_id", ({ books }) => {
     const mock_book_title = "The Mocking Shard";
     const mock_author_id = "1234";
